# openapi/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Limiq.io API",
    "summary": "Identity & Permission Layer for Autonomous Agents",
    "description": "\n## Limiq.io API\n\nIdentity & permission layer for autonomous agents.\n\n### Auth model (MVP)\nSensitive routes require `X-Workspace-Id` and `X-Workspace-Key`. The key is returned\nonce by workspace bootstrap and is verified in constant time. The request\n`workspace_id` must match the authenticated workspace.\n\n`POST /workspaces` is a bootstrap route protected by `X-Bootstrap-Token`.\n\n### Error format\nBusiness errors are returned as:\n\n```json\n{\"detail\": {\"code\": \"SOME_CODE\", \"message\": \"Human readable message\"}}\n```\n\n### Main flow\n1. Create agent\n2. Create policy\n3. Bind policy to agent\n4. Request capability\n5. Verify signed action\n6. Query/export audit logs\n7. Check audit chain integrity\n",
    "termsOfService": "https://github.com/qurveai/limiq.io",
    "contact": {
      "name": "Limiq.io Maintainers",
      "url": "https://github.com/qurveai/limiq.io"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.5.0",
    "x-logo": {
      "url": "https://raw.githubusercontent.com/redocly/redoc/main/demo/logo.png",
      "altText": "Limiq.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/workspaces": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "description": "Creates a workspace for local bootstrap/dev flows.",
        "operationId": "create_workspace_endpoint_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceCreatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Workspace slug already exists.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_SLUG_ALREADY_EXISTS",
                    "message": "Workspace slug already exists"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Workspace bootstrap endpoint disabled by configuration.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_BOOTSTRAP_DISABLED",
                    "message": "Workspace bootstrap is disabled; configure LIMIQ_WORKSPACE_BOOTSTRAP_TOKEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace",
        "description": "Fetches a workspace by id within authenticated tenant context.",
        "operationId": "get_workspace_endpoint_workspaces__workspace_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_NOT_FOUND",
                    "message": "Workspace not found"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/policies": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Create Policy",
        "description": "Creates a versioned policy in a workspace.",
        "operationId": "create_policy_endpoint_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Policy version already exists.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "POLICY_VERSION_ALREADY_EXISTS",
                    "message": "Policy version already exists"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/agents/{agent_id}/bind_policy": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Bind Policy To Agent",
        "description": "Binds a policy to an agent, deactivating previous active binding if any.",
        "operationId": "bind_policy_endpoint_agents__agent_id__bind_policy_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyBindRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyBindingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AUTH_WORKSPACE_KEY_MISSING",
                    "message": "Missing X-Workspace-Key header"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Workspace mismatch with authenticated context.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "WORKSPACE_MISMATCH",
                    "message": "Workspace does not match authenticated context"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error on payload/query params.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "VALIDATION_ERROR",
                    "message": "Query param 'from' must be <= 'to'"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Agent or policy not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "POLICY_OR_AGENT_NOT_FOUND",
                    "message": "Agent or policy not found"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Agent is revoked and cannot be bound.",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "AGENT_REVOKED",
                    "message": "Agent is revoked"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PolicyBindRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id",
            "description": "Workspace identifier (must match X-Workspace-Id)."
          },
          "policy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Id",
            "description": "Policy id to bind."
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "policy_id"
        ],
        "title": "PolicyBindRequest"
      },
      "PolicyBindingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "policy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "bound_at": {
            "type": "string",
            "format": "date-time",
            "title": "Bound At"
          },
          "unbound_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unbound At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "agent_id",
          "policy_id",
          "status",
          "bound_at",
          "unbound_at"
        ],
        "title": "PolicyBindingResponse"
      },
      "PolicyCreateRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id",
            "description": "Workspace identifier (must match X-Workspace-Id)."
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Policy name."
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "Policy version, unique per (workspace, name)."
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Schema Version",
            "description": "Policy schema version.",
            "default": 1
          },
          "policy_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy Json",
            "description": "Policy payload."
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "name",
          "version",
          "policy_json"
        ],
        "title": "PolicyCreateRequest"
      },
      "PolicyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "policy_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "name",
          "version",
          "is_active",
          "schema_version",
          "policy_json",
          "created_at"
        ],
        "title": "PolicyResponse"
      },
...
      "WorkspaceCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Name",
            "description": "Human readable workspace name."
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug",
            "description": "Optional unique slug. Lowercase letters, numbers and dashes."
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "WorkspaceCreateRequest"
      },
      "WorkspaceCreatedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "Workspace API key. Returned once during bootstrap."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status",
          "created_at",
          "api_key"
        ],
        "title": "WorkspaceCreatedResponse"
      },
      "WorkspaceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status",
          "created_at"
        ],
        "title": "WorkspaceResponse"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Workspace-Key"
      }
    }
  },
...
}